requirements.txt
kafka-python
langchain
langchain-google-genai
google-generativeai
pypdf2
streamlit
python-dotenv